+++ http4k-incubator/src/test/kotlin/org/http4k/tracing/TracerBulletTest.kt
@@ -3,7 +3,6 @@ package org.http4k.tracing
 import org.http4k.core.Response
 import org.http4k.core.Status.Companion.OK
 import org.http4k.core.with
-import org.http4k.events.MetadataEvent
 import org.http4k.events.then
 import org.http4k.format.Jackson
 import org.http4k.routing.reverseProxy
@@ -45,11 +44,10 @@ class TracerBulletTest {
             bodyString.isEqualTo("bob")
         }
 
-        approver.assertApproved(Response(OK).with(Jackson.autoBody<Any>().toLens() of
-            TracerBullet(HttpTracer(actorFrom), MyCustomTracer(actorFrom))(recording.toList())))
-        expectThat(
-            TracerBullet(HttpTracer(actorFrom), MyCustomTracer(actorFrom))(recording.toList())
-        ).isEqualTo(listOf(expectedCallTree))
+        val traces = TracerBullet(HttpTracer(actorFrom), DbTracer(actorFrom))(recording.toList())
+
+        approver.assertApproved(Response(OK).with(Jackson.autoBody<Any>().toLens() of traces))
+        expectThat(traces).isEqualTo(listOf(expectedCallTree))
     }
 }
 
@@ -76,12 +74,14 @@ val expectedCallTree = RequestResponse(
                     listOf()
                 )
             )
-        ), BiDirectional(
+        ),
+        BiDirectional(
             Actor("EntryPoint", System),
             Actor("db", Database),
             "EntryPoint",
             listOf()
-        ), RequestResponse(
+        ),
+        RequestResponse(
             Actor("EntryPoint", System),
             Actor("Child2", System),
             "GET {name}",